requirements.txt
# requests==2.28.1
Faker==15.3.4
Flask==2.2.2
pymongo==4.0.1
kafka-python==2.0.2
mysql-connector-python==8.0.31
neo4j==5.3.0